% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabMap.R
\name{addConceptsToDict}
\alias{addConceptsToDict}
\title{Function to add concept IDs/names/domains to a results dictionary}
\usage{
addConceptsToDict(resDict, cdm)
}
\arguments{
\item{resDict}{A dictionary of keywords created via makeDict}

\item{cdm}{a cdm_reference object created using the CDMConnector package}
}
\description{
Function to add concept IDs/names/domains to a results dictionary
}
