% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSummary.R
\name{addDictToRes}
\alias{addDictToRes}
\title{Add dictionary information directly to the res file}
\usage{
addDictToRes(res, conceptDict)
}
\arguments{
\item{res}{a cdm_reference object created using the CDMConnector package}

\item{conceptDict}{A dictionary of keywords created via makeDict with
attached OMOP concept name and ID information}
}
\description{
Add dictionary information directly to the res file
}
