% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{as_num}
\alias{as_num}
\title{Large number syntax divider}
\usage{
as_num(x)
}
\arguments{
\item{x}{A vector of strings of numerics with punctuation marks}
}
\value{
A numeric vector.
}
\description{
Large number syntax divider
}
\examples{
as_num(c("123_456.78", "1-000-000"))
}
