% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabMap.R
\name{atc2Map}
\alias{atc2Map}
\title{Function that maps drug codes to atc2 class}
\usage{
atc2Map(resDict, cdm)
}
\arguments{
\item{resDict}{A dictionary of keywords created via makeDict}

\item{cdm}{a cdm_reference object created using the CDMConnector package}
}
\description{
Function that maps drug codes to atc2 class
}
