% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{base_26_calc}
\alias{base_26_calc}
\title{Convert an integer to base 26}
\usage{
base_26_calc(x)
}
\arguments{
\item{x}{The integer to convert to base 26.}
}
\value{
A string of \code{x} in base 26.
}
\description{
Convert an integer to base 26
}
