% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{build_histogram_plot}
\alias{build_histogram_plot}
\title{Helper function to build histogram plot (internal use)}
\usage{
build_histogram_plot(df, title = "", show.legend = TRUE)
}
\arguments{
\item{df}{Dataframe with features and weights}

\item{title}{Plot title}

\item{show.legend}{Logical to show legend}
}
\description{
Helper function to build histogram plot (internal use)
}
\keyword{internal}
