% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{closest}
\alias{closest}
\title{Find the closest match in another vector}
\usage{
closest(x, table)
}
\arguments{
\item{x}{The element vector to look up.}

\item{table}{The vector of elements in which to find the closest match.}
}
\value{
The element from \code{table} that is closest in least absolute terms in \code{x}.
}
\description{
Find the closest match in another vector
}
