% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchStrategy.R
\name{combineSearchStrategy}
\alias{combineSearchStrategy}
\title{Function that build the pubmed search strategy}
\usage{
combineSearchStrategy(...)
}
\arguments{
\item{...}{a list of different field tags to add to the search}
}
\description{
Function that build the pubmed search strategy
}
