% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiHelpers.R
\name{commonKeywords}
\alias{commonKeywords}
\title{Function that lists common MeSH terms that are unlikely to be useful}
\usage{
commonKeywords()
}
\description{
Function that lists common MeSH terms that are unlikely to be useful
}
