% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp2Moy.R
\name{Comp2Moy}
\alias{Comp2Moy}
\title{Comp2Moy}
\usage{
Comp2Moy(Measure1, Measure2)
}
\arguments{
\item{Measure1}{a measure from a dataset}

\item{Measure2}{the same measure from a different dataset}
}
\description{
Compare 2 dataset using the right test according to normality and variances
}
\examples{

Comp2Moy(iris$Sepal.Length,iris$Petal.Length)

}
