% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSummary.R
\name{cumuDate}
\alias{cumuDate}
\title{Creates a cumulative sum matrix with dates as rows and key words as columns}
\usage{
cumuDate(res)
}
\arguments{
\item{res}{a cdm_reference object created using the CDMConnector package}
}
\value{
cumuDate A matrix of the cumulative sum of key words by date
}
\description{
Creates a cumulative sum matrix with dates as rows and key words as columns
}
