% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSummary.R
\name{cumuDate_Journal}
\alias{cumuDate_Journal}
\title{Creates a cumulative sum matrix with dates as rows and journals as columns}
\usage{
cumuDate_Journal(res)
}
\arguments{
\item{res}{a cdm_reference object created using the CDMConnector package}
}
\value{
cumuDate A matrix of the cumulative sum of journals by date
}
\description{
Creates a cumulative sum matrix with dates as rows and journals as columns
}
