% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchStrategy.R
\name{dateRange}
\alias{dateRange}
\title{Function to add a date range to the search query}
\usage{
dateRange(start, end)
}
\arguments{
\item{start}{the start date, must have format 'yyyy/mm/dd' as a character string}

\item{end}{the end date, must have format 'yyyy/mm/dd' as a character string}
}
\description{
Function to add a date range to the search query
}
