% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{dice_roller}
\alias{dice_roller}
\title{Dice roll simulator}
\usage{
dice_roller(sides = 6, times = 1, dice = 1)
}
\arguments{
\item{sides}{An integer on each of the dice.}

\item{times}{An integer on how many times to roll the dice.}

\item{dice}{An integer on the number of dice to roll}
}
\value{
A list of vectors for each of the dice that have been rolled.
}
\description{
Dice roll simulator
}
