% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{digital_sum}
\alias{digital_sum}
\title{Digital sum of a number}
\usage{
digital_sum(n)
}
\arguments{
\item{n}{The number to calculate the digital sum of.}
}
\description{
Digital sum of a number
}
