% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_features_one.R
\name{divide_features_one}
\alias{divide_features_one}
\title{Dividing features to create vectors with signal in the first omic for single data}
\usage{
divide_features_one(n_features_one, num.factor)
}
\arguments{
\item{n_features_one}{number of features of first omic}

\item{num.factor}{number of factor = '1'}
}
\description{
Dividing features to create vectors with signal in the first omic for single data
}
