% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_features_two.R
\name{divide_features_two}
\alias{divide_features_two}
\title{Dividing features to create vectors with signal in the second omic for single data}
\usage{
divide_features_two(n_features_two, num.factor)
}
\arguments{
\item{n_features_two}{number of features of first omic}

\item{num.factor}{number of factors (should be set to '1')}
}
\value{
A list of numeric vectors. Each vector contains 80\% of the features from one segment of the original feature set.
The number of segments is determined by the number of factors provided (\code{num.factor}).
\itemize{
\item Each vector is a subset of the original feature set, selected randomly to contain 80\% of the elements from the segment.
}
If the minimum segment size constraint is too large for the given feature length and number of segments, the function retries using the \code{divide_vector()} function.
}
\description{
Dividing features to create vectors with signal in the second omic for single data
}
