% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{excel_date}
\alias{excel_date}
\title{Convert excel number to a date}
\usage{
excel_date(date)
}
\arguments{
\item{date}{The numeric excel code to convert to a date object.}
}
\value{
A vector of date type.
}
\description{
Convert excel number to a date
}
