% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{excel_datetime}
\alias{excel_datetime}
\title{Convert excel number to a datetime}
\usage{
excel_datetime(date)
}
\arguments{
\item{date}{The numeric excel code to convert to a datetime object.}
}
\value{
A vector of datetime type.
}
\description{
Convert excel number to a datetime
}
