% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSummary.R
\name{exportBib}
\alias{exportBib}
\title{Outputs a bib file containing all result BibTeXs, derived via doi}
\usage{
exportBib(res, outfile = "bibs.bib")
}
\arguments{
\item{res}{a cdm_reference object created using the CDMConnector package}

\item{outfile}{a string file name to store bib output}
}
\description{
Outputs a bib file containing all result BibTeXs, derived via doi
}
