% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{factors}
\alias{factors}
\title{Factors of a number}
\usage{
factors(x, include_original = FALSE)
}
\arguments{
\item{x}{The integer to find the factors of.}

\item{include_original}{Should \code{x} by contained in the output?}
}
\value{
A vector of integers of the factors of \code{x}.
}
\description{
Factors of a number
}
