% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{fetchPubmed}
\alias{fetchPubmed}
\title{Fetch the results of a pubmed search}
\usage{
fetchPubmed(hits, searchStrategy)
}
\arguments{
\item{hits}{the ouptu of searchPubmed or an esearch list from the 'rentrez' package}

\item{searchStrategy}{a search strategy built using sumo}
}
\description{
Fetch the results of a pubmed search
}
