% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{\%&\%}
\alias{\%&\%}
\title{Binary string concatenation}
\usage{
x \%&\% y
}
\arguments{
\item{x}{The left-hand string to paste.}

\item{y}{The right-hand string to paste.}
}
\description{
Binary string concatenation
}
