% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{\%fmt\%}
\alias{\%fmt\%}
\title{Left-right string interpolation with formatting options of sprintf}
\usage{
x \%fmt\% y
}
\arguments{
\item{x}{The string to interpolate into.}

\item{y}{A list to interpolate into \code{x}}
}
\value{
The interpolated \code{x}.
}
\description{
Left-right string interpolation with formatting options of sprintf
}
