% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_expressions.R
\name{\%like\%}
\alias{\%like\%}
\title{SQL-like like wildcard}
\usage{
x \%like\% pattern
}
\arguments{
\item{x}{The item to match}

\item{pattern}{The matching regex pattern}
}
\value{
A vector of logicals.
}
\description{
SQL-like like wildcard
}
