% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{\%not in\%}
\alias{\%not in\%}
\title{Opposite of \code{in}}
\usage{
x \%not in\% table
}
\arguments{
\item{x}{The element to match.}

\item{table}{The set to match in.}
}
\value{
A logical vector showing whether \code{x} is not in \code{table}.
}
\description{
Opposite of \code{in}
}
