% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_expressions.R
\name{\%notlike\%}
\alias{\%notlike\%}
\title{SQL-like not-like wildcard}
\usage{
x \%notlike\% pattern
}
\arguments{
\item{x}{The item to match}

\item{pattern}{The matching regex pattern}
}
\value{
A vector of logicals.
}
\description{
SQL-like not-like wildcard
}
