% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{\%swap\%}
\alias{\%swap\%}
\title{Infix operator to swap the values of variables}
\usage{
x \%swap\% y
}
\arguments{
\item{x}{The first element to swap}

\item{y}{The second element to swap}
}
\description{
Infix operator to swap the values of variables
}
