% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsumo.R
\name{graphsumo}
\alias{graphsumo}
\title{graphsumo}
\usage{
graphsumo(dataset, abs, ord, MFV = "vir", alphajitter = 1, n)
}
\arguments{
\item{dataset}{the dataset organized as col = cond, Name, value and 1 row = 1 obs from 1 sample}

\item{abs}{parameter to be plotted in x ex : genotype}

\item{ord}{parameter to be plotted in y ex : value}

\item{MFV}{can be vir for viridis scale, male for b&w or female for red&white}

\item{n}{number of populations}

\item{alpha}{transparence}
}
\description{
plot a nice and consistent graph
}
\examples{

  graphsumo(iris, iris$Species, iris$Petal.Length, "vir")

}
