% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{group_if_consecutive}
\alias{group_if_consecutive}
\title{Create groups if values are consecutive}
\usage{
group_if_consecutive(x)
}
\arguments{
\item{x}{The vector over which to create groups.}
}
\value{
A integer vector showing the consecutive groups of \code{x}.
}
\description{
Create groups if values are consecutive
}
