% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{GSEA}
\usage{
GSEA(gene_list, GO_file, pval)
}
\arguments{
\item{gene_list}{list of gene /w genename and foldchange}

\item{GO_file}{List of gene of interest}

\item{pval}{p value}
}
\description{
gives enrichment
}
\examples{



}
