% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiHelpers.R
\name{highImpactJournals}
\alias{highImpactJournals}
\title{Function that lists high impact journals for a search}
\usage{
highImpactJournals()
}
\description{
Function that lists high impact journals for a search
}
