% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabMap.R
\name{icd10Map}
\alias{icd10Map}
\title{Function that maps condition codes to icd 10 chapters}
\usage{
icd10Map(resDict, cdm)
}
\arguments{
\item{resDict}{A dictionary of keywords created via makeDict}

\item{cdm}{a cdm_reference object created using the CDMConnector package}
}
\description{
Function that maps condition codes to icd 10 chapters
}
