% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchStrategy.R
\name{journalSearch}
\alias{journalSearch}
\title{Function to build a search query based on journals}
\usage{
journalSearch(journals)
}
\arguments{
\item{journals}{a character vector of journals to use in the search}
}
\description{
Function to build a search query based on journals
}
