% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchStrategy.R
\name{keywordSearch}
\alias{keywordSearch}
\title{Function to build a search query based on keywords}
\usage{
keywordSearch(keyword)
}
\arguments{
\item{keyword}{a character vector of keywords to use in the search}
}
\description{
Function to build a search query based on keywords
}
