% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{len_uni}
\alias{len_uni}
\title{Number of unique elements in the vector}
\usage{
len_uni(x)
}
\arguments{
\item{x}{The vector to find the number of unique elements over.}
}
\value{
An integer of the number of unique elements in \code{x}.
}
\description{
Number of unique elements in the vector
}
