% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{make_project}
\usage{
make_project(path, ref = TRUE, IHC = FALSE)
}
\arguments{
\item{path}{= path}

\item{IHC}{true if it's an IHC and I want jpg, czi and QP folders}
}
\description{
Strongly inspired by "https://github.com/nibortolum/manageproject". but more specific to my style of coding
}
\examples{


}
