% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{nan_to_na}
\alias{nan_to_na}
\title{Convert \code{NaN} to \code{NA} in vectors}
\usage{
nan_to_na(x, na_type = NULL)
}
\arguments{
\item{x}{The vector in which to replace \code{NaN} values.}

\item{na_type}{The NA type to replace \code{NaN} values with, eg: \code{NA}; \code{NA_real_}; etc.}
}
\value{
A vector that has \code{NA} values replaced..
}
\description{
Convert \code{NaN} to \code{NA} in vectors
}
