% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultPlots.R
\name{plotConditions}
\alias{plotConditions}
\title{Plot condition-specific OMOP results for a given results object}
\usage{
plotConditions(res, conceptDict)
}
\arguments{
\item{res}{a cdm_reference object created using the CDMConnector package}

\item{conceptDict}{A dictionary of keywords created via makeDict with
attached OMOP concept name and ID information}
}
\description{
Plot condition-specific OMOP results for a given results object
}
