% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSummary.R
\name{plotKeywords}
\alias{plotKeywords}
\title{Function to generate summary plots of key_words from a dict object which
contains MeSH terms mapped to OMOP concept codes via SnoMED and RxNorm}
\usage{
plotKeywords(res)
}
\arguments{
\item{res}{A fetched sumo object}
}
\description{
Function to generate summary plots of key_words from a dict object which
contains MeSH terms mapped to OMOP concept codes via SnoMED and RxNorm
}
