% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultPlots.R
\name{plotMap}
\alias{plotMap}
\title{Plot a map of the country of origin MeSH term}
\usage{
plotMap(conceptDict)
}
\arguments{
\item{conceptDict}{A dictionary of keywords created via makeDict with}
}
\description{
Plot a map of the country of origin MeSH term
}
