% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultPlots.R
\name{plotRollUp}
\alias{plotRollUp}
\title{Plot a bar chart of rollup terms}
\usage{
plotRollUp(conceptDict, N = 10)
}
\arguments{
\item{conceptDict}{A dictionary of keywords created via makeDict}
}
\description{
Plot a bar chart of rollup terms
}
