% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{polygon_area}
\alias{polygon_area}
\title{Area of a polyon new}
\usage{
polygon_area(x, y)
}
\arguments{
\item{x}{The x-coordinates of the polygon.}

\item{y}{The y-coordinates of the polygon.}
}
\description{
Area of a polyon new
}
