% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{proportion}
\alias{proportion}
\title{Evaluate the proportion based on a grouped variable}
\usage{
proportion(x, na_rm = TRUE)
}
\arguments{
\item{x}{The vector to find the element-wise proportion}

\item{na_rm}{Remove NA values before evaluating the proportion?}
}
\value{
A numeric vector that shows the element-wise proportion of \code{x}.
}
\description{
Evaluate the proportion based on a grouped variable
}
