% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchStrategy.R
\name{removeKeywords}
\alias{removeKeywords}
\title{Function that removes keywords from the pubmed search strategy}
\usage{
removeKeywords(searchStrategy)
}
\arguments{
\item{searchStrategy}{A completed search strategy generated by combineSearchStrategy}
}
\description{
Function that removes keywords from the pubmed search strategy
}
