% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{replace_after}
\alias{replace_after}
\title{Replace all text after a given character}
\usage{
replace_after(string, symbol, replacement)
}
\arguments{
\item{string}{The string to replace in.}

\item{symbol}{The symbol to find in \code{string}}

\item{replacement}{The replacement of \code{string} before \code{symbol}.}
}
\description{
Replace all text after a given character
}
