% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{replace_after_old}
\alias{replace_after_old}
\title{Function to create regex to gsub replacement after a symbol}
\usage{
replace_after_old(
  string,
  symbol,
  replacement,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{string}{The string to replace in.}

\item{symbol}{The symbol to find in \code{string}}

\item{replacement}{The replacement of \code{string} after \code{symbol}.}

\item{ignore.case}{If FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}

\item{perl}{Logical. Should Perl-compatible regexps be used?}

\item{fixed}{Logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}

\item{useBytes}{Logical. If TRUE the matching is done byte-by-byte rather than character-by-character. See ‘Details’.}
}
\description{
Function to create regex to gsub replacement after a symbol
}
