% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{round_down_to_nearest}
\alias{round_down_to_nearest}
\title{Round down}
\usage{
round_down_to_nearest(x, n)
}
\arguments{
\item{x}{The value to round down.}

\item{n}{The base to round down to.}
}
\value{
An integer of the rounded down \code{x} to the nearest \code{n}.
}
\description{
Round down
}
