% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{round_to_nearest}
\alias{round_to_nearest}
\title{Round to nearest}
\usage{
round_to_nearest(x, n)
}
\arguments{
\item{x}{The value to round.}

\item{n}{The base to round to.}
}
\value{
An integer of the rounded \code{x} to the nearest \code{n}.
}
\description{
Round to nearest
}
