% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{round_up_to_nearest}
\alias{round_up_to_nearest}
\title{Round up}
\usage{
round_up_to_nearest(x, n)
}
\arguments{
\item{x}{The value to round up.}

\item{n}{The base to round up to.}
}
\value{
An integer of the rounded up \code{x} to the nearest \code{n}.
}
\description{
Round up
}
