% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pheatmap.R
\name{save_pheatmap}
\alias{save_pheatmap}
\title{save_pheatmap}
\usage{
save_pheatmap(x, filename, width = 480, height = 960)
}
\arguments{
\item{x}{heatmap name}

\item{filename}{directory and name with extension (pdf or png)}

\item{width, height}{the width and height of the graphics region in inches}
}
\description{
save pheatmap in pdf or png
}
\examples{

  \dontrun{
  HM <- pheatmap(iris[,-5])
  save_pheatmap(HM, "path/Heatmap.png")
           }

}
