% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{searchPubmed}
\alias{searchPubmed}
\title{Search pubmed using defined search strategy}
\usage{
searchPubmed(searchStrategy)
}
\arguments{
\item{searchStrategy}{a search strategy built using sumo}
}
\value{
an esearch list from the 'rentrez' package specifying the pmids resulting
from the search and count of hits
}
\description{
Search pubmed using defined search strategy
}
