% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceAll.R
\name{SourceAll}
\alias{SourceAll}
\title{SourceAll}
\usage{
SourceAll()
}
\description{
will source every functions located in the Sources folder if it contains functions
will better work with a project created with make_project function
}
