% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{sum_na}
\alias{sum_na}
\title{Count the number of NA in an object}
\usage{
sum_na(x)
}
\arguments{
\item{x}{The vector the calculate the number of NA elements}
}
\value{
An integer of the number of \code{NA} values in \code{x}.
}
\description{
Count the number of NA in an object
}
