% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumo_py.R
\name{sumo_mofa_backend}
\alias{sumo_mofa_backend}
\title{Detect and configure the MOFA2 backend for SUMO}
\usage{
sumo_mofa_backend()
}
\value{
A list with elements:
\itemize{
\item \code{method}: one of \code{"basilisk"}, \code{"reticulate"}, or \code{"pretrained"}.
\item \code{use_basilisk}: logical; \code{TRUE} if using basilisk.
\item \code{envname}: \code{NULL} or the name of a \code{conda} environment.
}
}
\description{
Detect and configure the MOFA2 backend for SUMO
}
\keyword{internal}
