% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretrained.R
\name{sumo_pretrained_mofa_available}
\alias{sumo_pretrained_mofa_available}
\title{List pretrained MOFA models included with SUMO}
\usage{
sumo_pretrained_mofa_available()
}
\value{
Character vector of file names present in inst/extdata (empty if none).
}
\description{
List pretrained MOFA models included with SUMO
}
