% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretrained.R
\name{sumo_pretrained_mofa_path}
\alias{sumo_pretrained_mofa_path}
\title{Path to a pretrained MOFA model shipped with SUMO}
\usage{
sumo_pretrained_mofa_path(which = c("SUMO", "CLL"))
}
\arguments{
\item{which}{One of "SUMO" or "CLL".}
}
\value{
Full file path to the pretrained model.
}
\description{
Path to a pretrained MOFA model shipped with SUMO
}
