% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchStrategy.R
\name{textSearch}
\alias{textSearch}
\title{Function to build a search query based on text}
\usage{
textSearch(text)
}
\arguments{
\item{text}{a character vector of text to use in the search}
}
\description{
Function to build a search query based on text
}
