% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchStrategy.R
\name{tiabSearch}
\alias{tiabSearch}
\title{Function to build a search query based on terms in the title and abstract}
\usage{
tiabSearch(tiab)
}
\arguments{
\item{mesh}{a character vector of terms in the title and abstract to use in the search}
}
\description{
Function to build a search query based on terms in the title and abstract
}
