% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triQP.R
\name{triQP}
\alias{triQP}
\title{triQP}
\usage{
triQP(df)
}
\arguments{
\item{df}{dataset extracted from a QPath analysis and where annotation correspond to "Genotype Sex"}
}
\description{
will create 2 columns containing respectively the genotype and the sex of the samples
}
\examples{


}
