% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{which_na}
\alias{which_na}
\title{Show the subset of \code{NA} in an object}
\usage{
which_na(x)
}
\arguments{
\item{x}{The vector to identify \code{NA} elements in}
}
\value{
A integer vector of indices of \code{x} with \code{NA} elements
}
\description{
Show the subset of \code{NA} in an object
}
