% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{base_26_calc}
\alias{base_26_calc}
\title{Convert an base 26 string to an integer}
\usage{
base_26_calc(x)
}
\arguments{
\item{x}{The base 26 string to convert to base 10}
}
\value{
An integer  of \code{x} in base 10.
}
\description{
Convert an base 26 string to an integer
}
