% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp3Moy.R
\name{Comp3Moy}
\alias{Comp3Moy}
\title{Comp3Moy}
\usage{
Comp3Moy(Condition, Measure, n = 3)
}
\arguments{
\item{Condition}{the different conditions named in the same column}

\item{Measure}{the values measured for every sample}

\item{n}{is the number or replicates per condition}
}
\description{
Compare 2 or more datasets using the right test according to normality and variances
Condition and Measure should be different columns of the same dataset (Dataset$colA)
@import rstatix
@import dplyr
@examples

Comp3Moy(iris$Species, iris$Sepal.Length, n=150)
}
