% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{digital_sum}
\alias{digital_sum}
\title{Iteratively calculate the digital sum of a number}
\usage{
digital_sum(x)
}
\arguments{
\item{x}{The integer to calculate the digital sum of}
}
\value{
An integer of the digitical sum of \code{x}.
}
\description{
Iteratively calculate the digital sum of a number
}
