% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_features_two.R
\name{divide_features_two}
\alias{divide_features_two}
\title{Dividing features to create vectors with signal in the second omic for single data}
\usage{
divide_features_two(n_features_two, num.factor)
}
\arguments{
\item{n_features_two}{number of features of second omic}

\item{num.factor}{type of factors - single or multiple}
}
\description{
Dividing features to create vectors with signal in the second omic for single data
}
