% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_samples.R
\name{divide_samples}
\alias{divide_samples}
\title{Global Variable}
\usage{
divide_samples(n_samples, num, min_size)
}
\arguments{
\item{n_samples}{number of samples}

\item{num}{number of factors}

\item{min_size}{Minimum length of any samples scores}
}
\description{
A global variable used in multiple functions.
}
