% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_vector.R
\name{divide_vector}
\alias{divide_vector}
\title{#' Global Variable
#'
#' A global variable used in multiple functions.
#'
#'}
\usage{
divide_vector(n_samples, num, min_size)
}
\arguments{
\item{n_samples}{number of samples}

\item{num}{number of factors}

\item{min_size}{Minimum length of any samples scores

#' ## ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ Updated IN USE (IN USE): Simulate the samples scores  (IN USE) ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~}
}
\description{
#' Global Variable
#'
#' A global variable used in multiple functions.
#'
#'
}
