% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{factors}
\alias{factors}
\title{Factors of a number new}
\usage{
factors(input, include_original)
}
\arguments{
\item{input}{The number to calculate the factors of.}
}
\description{
Factors of a number new
}
