% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_one.R
\name{feature_selection_one}
\alias{feature_selection_one}
\title{Dividing features to create vectors with signal in the first omic}
\usage{
feature_selection_one(n_features_one, num.factor, no_factor)
}
\arguments{
\item{n_features_one}{number of features of first omic}

\item{num.factor}{type of factors - single or multiple}

\item{no_factor}{number of factors}
}
\description{
Dividing features to create vectors with signal in the first omic
}
