% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_two.R
\name{feature_selection_two}
\alias{feature_selection_two}
\title{Dividing features to create vectors with signal in the second omic}
\usage{
feature_selection_two(n_features_two, num.factor, no_factor)
}
\arguments{
\item{n_features_two}{number of features of second omic}

\item{num.factor}{type of factors - single or multiple}

\item{no_factor}{number of factors}
}
\description{
Dividing features to create vectors with signal in the second omic
}
