% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{len_uni}
\alias{len_uni}
\title{I'm sick of typing length(unique(x))}
\usage{
len_uni(x)
}
\arguments{
\item{x}{The vector to find the number of unique elements over.}
}
\value{
An integer of the number of unique elements in \code{x}.
}
\description{
I'm sick of typing length(unique(x))
}
