% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\alias{make_project2}
\title{make_project}
\usage{
make_project2(path, IHC = FALSE, path_to_QuPath)
}
\arguments{
\item{path}{path where to create project}

\item{IHC}{true if it's an IHC and I want tiff, czi and QP folders}

\item{path_to_QuPath}{the path where the QuPath (console).exe is located, should look like this on windows "C:/Users/user_name/AppData/Local/QuPath-0.4.1/"}
}
\description{
Strongly inspired by "https://github.com/nibortolum/manageproject". but contains extra specific features
IHC=T provides folder to store raw and cropped pictures and create a QuPath project for further analyses
Detection of current directory enables more flexibility
SourceAll function makes things easier when a lot of external functions are used
}
