% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSummary.R
\name{makeDict}
\alias{makeDict}
\title{Function to build a dictionary of all the MeSH terms in a result
Prototype function used before MeSH mapping is complete}
\usage{
makeDict(res, cdm, removeCommon = TRUE, rollup = TRUE)
}
\arguments{
\item{res}{A fetched sumo object containing PMIDs and key words}

\item{cdm}{a cdm_reference object created using the CDMConnector package}

\item{removeCommon}{A toggle on whether or not to remove the most common keywords}
}
\description{
Function to build a dictionary of all the MeSH terms in a result
Prototype function used before MeSH mapping is complete
}
