% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmixCraftHD.R
\name{OmixCraftHD}
\alias{OmixCraftHD}
\title{Simulation of high-dimensional data with predefined single factor or multiple factors in multi-omics}
\usage{
OmixCraftHD(
  vector_features = c(2000, 2000),
  n_samples = 50,
  sigmas_vector = c(3, 5),
  n_factors = 3,
  num.factor = "multiple",
  advanced_dist = NULL
)
}
\arguments{
\item{vector_features}{Vector of features assigned to the two simulated datasets respectively '1' first dataset, '2' second dataset}

\item{n_samples}{The number of samples common between the two simulated datasets}

\item{sigmas_vector}{Vector for the noise variability for the two simulated datasets respectively,  '1' first dataset, '2' second dataset}

\item{n_factors}{Number of predefined factors}

\item{num.factor}{Category of factors to be simulated specified as 'single', or 'multiple'.}

\item{advanced_dist}{Applicable only when num.factor = 'multiple'. Contains six possible arguments, '', NULL, 'mixed', 'omic.one', or 'omic.two', 'exclusive'}
}
\description{
Simulation of high-dimensional data with predefined single factor or multiple factors in multi-omics
}
\examples{
# Examples
set.seed(1234)
output_obj <- OmixCraftHD(
  vector_features = c(2000,3000),
  sigmas_vector=c(8,5),
  n_samples=100,
  n_factors=5,
  num.factor='multiple',
  advanced_dist='mixed'
)
output_obj <- OmixCraftHD(
  vector_features = c(5000,3000),
  sigmas_vector=c(3,4),
  n_samples=30, n_factors=1
)
}
