% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factor.R
\name{plot_factor}
\alias{plot_factor}
\title{Visualization of factor scores}
\usage{
plot_factor(sim_object = NULL, factor_num = NULL)
}
\arguments{
\item{sim_object}{R object containing data to be plotted}

\item{factor_num}{Factor to be plotted.}
}
\description{
Visualization of factor scores
}
