% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{plot_weights}
\alias{plot_weights}
\title{Visualizing the loading of the features}
\usage{
plot_weights(
  sim_object = NULL,
  factor_num = 1,
  data = "omic.one",
  type = "scatter"
)
}
\arguments{
\item{sim_object}{R object containing data to be plotted}

\item{factor_num}{Factor to be plotted.}

\item{data}{Section of the integrated data to be plotted, omic.one or omic.two are the options}

\item{type}{Type of plot. Scatter plot and histogram are the only allowed plots}
}
\description{
Visualizing the loading of the features
}
