% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{polygon_area}
\alias{polygon_area}
\title{Area of a given polygon}
\usage{
polygon_area(x, y)
}
\arguments{
\item{x}{Ordered vector of x coordinates}

\item{y}{Ordered vector of y coordinates}
}
\description{
Area of a given polygon
}
