% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSummary.R
\name{printAbstract}
\alias{printAbstract}
\title{Function to pretty print Abstracts from a res object}
\usage{
printAbstract(res)
}
\arguments{
\item{res}{A fetched sumo object}
}
\description{
Function to pretty print Abstracts from a res object
}
