% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{replace_before}
\alias{replace_before}
\title{Replace all text before a given character}
\usage{
replace_before(string, symbol, replacement)
}
\description{
Replace all text before a given character
}
