% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabMap.R
\name{searchOmop}
\alias{searchOmop}
\title{Function to build a search query based on omop concept ids}
\usage{
searchOmop(conceptIds, cdm)
}
\arguments{
\item{conceptIds}{a integer string of concept ids from OMOP vocabulary}

\item{cdm}{a cdm_reference object created using the CDMConnector package}
}
\description{
Function to build a search query based on omop concept ids
}
