% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumo.R
\docType{package}
\name{sumo-package}
\alias{sumo}
\alias{sumo-package}
\title{sumo: Pubmed extractor based on OMOP vocabulary}
\description{
Helps search through pubmed based on terms linked to the OMOP vocabulary.
}
\author{
\strong{Maintainer}: Laurence Lawrence-Archer \email{laurence.lawrence@odysseusinc.com}

Authors:
\itemize{
  \item Martin Lavallee \email{martin.lavallee@odysseusinc.com}
}

}
\keyword{internal}
